default from float to double

+++ include/cvplot/figure.h
@@ -12,15 +12,15 @@
 namespace cvplot {
 
 struct Point2 {
-  float x, y;
+  double x, y;
   Point2() : Point2(0, 0) {}
-  Point2(float x, float y) : x(x), y(y) {}
+  Point2(double x, double y) : x(x), y(y) {}
 };
 
 struct Point3 {
-  float x, y, z;
+  double x, y, z;
   Point3() : Point3(0, 0, 0) {}
-  Point3(float x, float y, float z) : x(x), y(y), z(z) {}
+  Point3(double x, double y, double z) : x(x), y(y), z(z) {}
 };
 
 enum Type {
@@ -52,41 +52,42 @@ class Series {
   auto color(Color color) -> Series &;
   auto dynamicColor(bool dynamic_color) -> Series &;
   auto legend(bool legend) -> Series &;
-  auto add(const std::vector<std::pair<float, float>> &data) -> Series &;
-  auto add(const std::vector<std::pair<float, Point2>> &data) -> Series &;
-  auto add(const std::vector<std::pair<float, Point3>> &data) -> Series &;
-  auto addValue(const std::vector<float> &values) -> Series &;
+  auto add(const std::vector<std::pair<double, double>> &data) -> Series &;
+  auto add(const std::vector<std::pair<double, Point2>> &data) -> Series &;
+  auto add(const std::vector<std::pair<double, Point3>> &data) -> Series &;
+  auto addValue(const std::vector<double> &values) -> Series &;
   auto addValue(const std::vector<Point2> &values) -> Series &;
   auto addValue(const std::vector<Point3> &values) -> Series &;
-  auto add(float key, float value) -> Series &;
-  auto add(float key, Point2 value) -> Series &;
-  auto add(float key, Point3 value) -> Series &;
-  auto addValue(float value) -> Series &;
-  auto addValue(float value_a, float value_b) -> Series &;
-  auto addValue(float value_a, float value_b, float value_c) -> Series &;
-  auto set(const std::vector<std::pair<float, float>> &data) -> Series &;
-  auto set(const std::vector<std::pair<float, Point2>> &data) -> Series &;
-  auto set(const std::vector<std::pair<float, Point3>> &data) -> Series &;
-  auto setValue(const std::vector<float> &values) -> Series &;
+  auto add(double key, double value) -> Series &;
+  auto add(double key, Point2 value) -> Series &;
+  auto add(double key, Point3 value) -> Series &;
+  auto addValue(double value) -> Series &;
+  auto addValue(double value_a, double value_b) -> Series &;
+  auto addValue(double value_a, double value_b, double value_c) -> Series &;
+  auto set(const std::vector<std::pair<double, double>> &data) -> Series &;
+  auto set(const std::vector<std::pair<double, Point2>> &data) -> Series &;
+  auto set(const std::vector<std::pair<double, Point3>> &data) -> Series &;
+  auto setValue(const std::vector<double> &values) -> Series &;
   auto setValue(const std::vector<Point2> &values) -> Series &;
   auto setValue(const std::vector<Point3> &values) -> Series &;
-  auto set(float key, float value) -> Series &;
-  auto set(float key, float value_a, float value_b) -> Series &;
-  auto set(float key, float value_a, float value_b, float value_c) -> Series &;
-  auto setValue(float value) -> Series &;
-  auto setValue(float value_a, float value_b) -> Series &;
-  auto setValue(float value_a, float value_b, float value_c) -> Series &;
+  auto set(double key, double value) -> Series &;
+  auto set(double key, double value_a, double value_b) -> Series &;
+  auto set(double key, double value_a, double value_b, double value_c)
+      -> Series &;
+  auto setValue(double value) -> Series &;
+  auto setValue(double value_a, double value_b) -> Series &;
+  auto setValue(double value_a, double value_b, double value_c) -> Series &;
   auto clear() -> Series &;
 
   auto label() const -> const std::string &;
   auto legend() const -> bool;
   auto color() const -> Color;
-  void draw(void *buffer, float x_min, float x_max, float y_min, float y_max,
-            float x_axis, float xs, float xd, float ys, float yd, float y_axis,
-            int unit, float offset) const;
+  void draw(void *buffer, double x_min, double x_max, double y_min,
+            double y_max, double x_axis, double xs, double xd, double ys,
+            double yd, double y_axis, int unit, double offset) const;
   auto collides() const -> bool;
   void dot(void *b, int x, int y, int r) const;
-  void bounds(float &x_min, float &x_max, float &y_min, float &y_max,
+  void bounds(double &x_min, double &x_max, double &y_min, double &y_max,
               int &n_max, int &p_max) const;
   void verifyParams() const;
 
@@ -96,7 +97,7 @@ class Series {
 
  protected:
   std::vector<int> entries_;
-  std::vector<float> data_;
+  std::vector<double> data_;
   enum Type type_;
   Color color_;
   std::string label_;
@@ -136,8 +137,8 @@ class Figure {
   auto subaxisColor() -> Color;
   auto textColor() -> Color;
 
-  void draw(void *buffer, float x_min, float x_max, float y_min, float y_max,
-            int n_max, int p_max) const;
+  void draw(void *buffer, double x_min, double x_max, double y_min,
+            double y_max, int n_max, int p_max) const;
   auto drawFit(void *buffer) const -> int;
   auto drawFile(const std::string &filename, Size size) const -> bool;
   void show(bool flush = true) const;
